+++ tests/e2e/test_graphrag_pipeline_e2e.py
@@ -18,7 +18,7 @@ from iris_vector_rag.core.connection import ConnectionManager
 from iris_vector_rag.core.models import Document
 from iris_vector_rag.pipelines.graphrag import GraphRAGPipeline, KnowledgeGraphNotPopulatedException
 from iris_vector_rag.storage.vector_store_iris import IRISVectorStore
-from common.utils import get_llm_func
+from iris_vector_rag.common.utils import get_llm_func
 
 
 @pytest.fixture(scope="function")