+++ tests/e2e/test_vector_store_comprehensive_e2e.py
@@ -10,7 +10,7 @@ from iris_vector_rag.config.manager import ConfigurationManager
 from iris_vector_rag.core.connection import ConnectionManager
 from iris_vector_rag.core.models import Document
 from iris_vector_rag.storage.vector_store_iris import IRISVectorStore
-from common.utils import get_embedding_func
+from iris_vector_rag.common.utils import get_embedding_func
 
 
 @pytest.fixture(scope="module")